fix: rename references to lib
