+++ tests/integration/manifests-tester/src/charm.py
@@ -12,7 +12,7 @@ from pathlib import Path
 from typing import List
 
 import yaml
-from charms.kubernetes_manifests.v0.kubernetes_manifests import (
+from charms.resource_dispatcher.v0.kubernetes_manifests import (
     KubernetesManifest,
     KubernetesManifestRequirerWrapper,
     KubernetesManifestsRequirer,
